requirements.txt
asgiref==3.5.2
Django==4.0.6
mccabe==0.6.1
psycopg2-binary==2.9.3
sqlparse==0.4.2
djangorestframework==3.13.1
pytz==2022.1
